NIFI-11137 package refactor
