+++ nifi-nar-bundles/nifi-jms-bundle/nifi-jms-processors/src/main/java/org/apache/nifi/jms/processors/ioconcept/reader/FlowFileReaderCallback.java
@@ -14,11 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.nifi.jms.processors.strategy.publisher;
+package org.apache.nifi.jms.processors.ioconcept.reader;
 
 import org.apache.nifi.flowfile.FlowFile;
-import org.apache.nifi.processor.ProcessSession;
 
-public interface FlowFileReader {
-    void read(ProcessSession session, FlowFile flowFile, MessageHandler messageHandler, FlowFileReaderCallback flowFileReaderCallback);
+public interface FlowFileReaderCallback {
+    void onSuccess(FlowFile flowFile, int processedRecords, boolean isRecover, long transmissionMillis);
+    void onFailure(FlowFile flowFile, int processedRecords, long transmissionMillis, Exception e);
 }
